+++ appObjects/FlatCAMDocument.py
@@ -66,10 +66,8 @@ class DocumentObject(FlatCAMObj):
         self.to_form()
 
         # Show/Hide Advanced Options
-        if self.app.defaults["global_app_level"] == 'b':
-            self.ui.level.setText('<span style="color:green;"><b>%s</b></span>' % _("Beginner"))
-        else:
-            self.ui.level.setText('<span style="color:red;"><b>%s</b></span>' % _("Advanced"))
+        app_mode = self.app.defaults["global_app_level"]
+        self.change_level(app_mode)
 
         self.document_editor_tab = AppTextEditor(app=self.app)
         stylesheet = """
@@ -91,6 +89,8 @@ class DocumentObject(FlatCAMObj):
         # ######################################################################
         # ######################## SIGNALS #####################################
         # ######################################################################
+        self.ui.level.toggled.connect(self.on_level_changed)
+
         self.document_editor_tab.buttonOpen.clicked.disconnect()
         self.document_editor_tab.buttonOpen.clicked.connect(lambda: self.document_editor_tab.handleOpen(filt=flt))
         self.document_editor_tab.buttonSave.clicked.disconnect()
@@ -171,6 +171,39 @@ class DocumentObject(FlatCAMObj):
 
         # Switch plot_area to CNCJob tab
         self.app.ui.plot_tab_area.setCurrentWidget(self.document_editor_tab)
+
+    def change_level(self, level):
+        """
+
+        :param level:   application level: either 'b' or 'a'
+        :type level:    str
+        :return:
+        """
+
+        if level == 'a':
+            self.ui.level.setChecked(True)
+        else:
+            self.ui.level.setChecked(False)
+        self.on_level_changed(self.ui.level.isChecked())
+
+    def on_level_changed(self, checked):
+        if not checked:
+            self.ui.level.setText('%s' % _('Beginner'))
+            self.ui.level.setStyleSheet("""
+                                                QToolButton
+                                                {
+                                                    color: green;
+                                                }
+                                                """)
+
+        else:
+            self.ui.level.setText('%s' % _('Advanced'))
+            self.ui.level.setStyleSheet("""
+                                                QToolButton
+                                                {
+                                                    color: red;
+                                                }
+                                                """)
 
     def on_autocomplete_changed(self, state):
         if state: